Refactor code in Commands to be more in line with coding conventions
