+++ src/main/java/mona/command/ByeCommand.java
@@ -1,14 +1,14 @@
-package command;
+package mona.command;
 
-import storage.Storage;
-import task.TaskList;
-import ui.Ui;
+import mona.storage.Storage;
+import mona.task.TaskList;
+import mona.ui.Ui;
 
 public class ByeCommand extends Command {
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
-        super.setExit();
+        super.markAsExit();
         ui.bye();
     }
 }
