+++ src/main/java/mona/command/UnmarkCommand.java
@@ -1,32 +1,33 @@
-package command;
+package mona.command;
 
-import exception.MonaException;
-import storage.Storage;
-import task.Task;
-import task.TaskList;
-import ui.Ui;
+import mona.exception.MonaException;
+import mona.storage.Storage;
+import mona.task.Task;
+import mona.task.TaskList;
+import mona.ui.Ui;
 
 public class UnmarkCommand extends Command {
-    private int index;
+    private int taskIndex;
 
     public UnmarkCommand(int index) {
         super();
-        this.index = index;
+        this.taskIndex = index;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws MonaException {
-        if (index < 0 || index >= tasks.getSize()) {
-            throw new MonaException.TaskNotFoundException(index + 1);
+        if (taskIndex < 0 || taskIndex >= tasks.getSize()) {
+            throw new MonaException.TaskNotFoundException(taskIndex + 1);
         }
 
-        Task task = tasks.getTask(index);
+        Task task = tasks.getTask(taskIndex);
 
         if (task.getIsDone()){
             task.markAsUndone();
         } else {
             throw new MonaException.TaskAlreadyUndoneException(task);
         }
+
         storage.saveData(tasks);
         ui.showUnmarkMessage(task);
     }
