+++ src/main/java/mona/task/Task.java
@@ -1,4 +1,4 @@
-package task;
+package mona.task;
 
 /**
  * Represents a task with a description and completion status.
