+++ src/main/java/mona/task/Todo.java
@@ -1,4 +1,4 @@
-package task;
+package mona.task;
 
 public class Todo extends Task {
