+++ src/test/java/mona/task/TodoTest.java
@@ -1,10 +1,12 @@
-package task;
+package mona.task;
 
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import mona.task.Todo;
 
 
 public class TodoTest {
